PropertyEditorConverter back to top-level class
